+++ src/Geo/Models/Currency.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Enraiged\Geo\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+
+class Currency extends Model
+{
+    /** @var  string  The database table name. */
+    protected $table = 'currencies';
+
+    /** @var  bool  Indicates if the model should be timestamped. */
+    public $timestamps = false;
+
+    /** @var  array  The attributes that aren't mass assignable. */
+    protected $guarded = [
+        'id',
+    ];
+
+    /** @var  array  The attributes that should be cast. */
+    protected $casts = [
+        'is_active' => 'boolean',
+    ];
+
+    /**
+     *  @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function countries(): HasMany
+    {
+        return $this->hasMany(Country::class, 'currency', 'code');
+    }
+}